--- test/units/modules/network/junos/fixtures/junos_command_show_version.txt
@@ -1,4 +1,0 @@
-Hostname: vsrx01
-Model: vSRX
-Junos: 15.1X49-D15.4
-JUNOS Software Release [15.1X49-D15.4]
